# perf: registerify xhash.RNG.Uint64