```
goos: linux
goarch: amd64
pkg: github.com/zephyrtronium/xirho/xmath
cpu: AMD Ryzen 5 3600 6-Core Processor
          │   /tmp/old   │           /tmp/new            │
          │    sec/op    │    sec/op     vs base         │
Uint64-12   2.270n ± 23%   1.922n ± 53%  ~ (p=1.000 n=7)

          │   /tmp/old    │            /tmp/new            │
          │      B/s      │      B/s       vs base         │
Uint64-12   3.282Gi ± 30%   3.876Gi ± 50%  ~ (p=1.000 n=7)
```

Unlike LLVM and GCC the go compiler is an extremely limited and extremely conservative understanding of memory side effects. This led to it compiling the memory operation more or less as-is because it's simple understanding lead it to think that maybe the various fields of rng aliases to each other.

Hoisting out to a variable like this allows it to quite easily lift it to SSA since the local variables are never taken as reference.

The old code used to operate directly on the memory fields:
```
MOVQ 0x10(SP), DX
LEAQ 0(DX)(DX*4), BX
ROLQ $0x7, BX
SHLQ $0x11, DX
MOVQ 0x8(SP), SI
XORQ 0x18(SP), SI
XORQ SI, DX
MOVQ SI, 0x18(SP)
MOVQ 0x10(SP), DI
XORQ 0x20(SP), DI
MOVQ DI, 0x20(SP)
XORQ SI, 0x10(SP)
XORQ DI, 0x8(SP)
MOVQ DX, 0x18(SP)
ROLQ $0x2d, DI
MOVQ DI, 0x20(SP)
```
While the new on lifts theses to registers:
```
MOVQ 0x8(SP), DX
MOVQ 0x10(SP), BX
MOVQ 0x18(SP), SI
MOVQ 0x20(SP), DI
LEAQ 0(BX)(BX*4), R8
ROLQ $0x7, R8
MOVQ BX, R9
SHLQ $0x11, BX
XORQ DX, SI
XORQ R9, DI
XORQ SI, R9
XORQ DI, DX
XORQ SI, BX
ROLQ $0x2d, DI
MOVQ DX, 0x8(SP)
MOVQ R9, 0x10(SP)
MOVQ BX, 0x18(SP)
MOVQ DI, 0x20(SP)
```